Revert "Revert "change to long""
This reverts commit 1f394b1901800f62a22333c3464dd32f8547d001.

+++ src/main/java/io/github/levtey/TokenAPI/Place.java
@@ -5,9 +5,9 @@ import java.util.UUID;
 public class Place {
 	
 	public final UUID uuid;
-	public final int amount;
+	public final long amount;
 	
-	public Place(UUID uuid, int amount) {
+	public Place(UUID uuid, long amount) {
 		this.uuid = uuid;
 		this.amount = amount;
 	}
